requirements.txt
streamlit>=1.28.0
ibflex>=0.15
pandas>=2.0.0
plotly>=5.15.0
pyyaml>=6.0
requests>=2.28.0
python-dateutil>=2.8.0
yfinance>=0.2.18 